src/test/java/com/promineotech/TestDemoUnitTest.java
package com.promineotech;

import static org.assertj.core.api.Assertions.assertThat;
import static org.assertj.core.api.Assertions.assertThatThrownBy;
import static org.junit.jupiter.api.Assertions.*;
import static org.junit.jupiter.params.provider.Arguments.arguments;
import static org.mockito.Mockito.doReturn;
import static org.mockito.Mockito.spy;

import java.util.stream.Stream;

import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;

class TestDemoUnitTest {
	
	private TestDemo testDemo;

	@BeforeEach
	void setUp() throws Exception {
		testDemo = new TestDemo();
	}

	@ParameterizedTest
	@MethodSource("com.promineotech.TestDemoUnitTest#argumentsForAddPositive")
	void assertThatTwoPositiveNumbersAreAddedCorrectly(int a, int b, int expected, boolean expectException) {
		if(!expectException) {
			assertThat(testDemo.addPositive(a, b)).isEqualTo(expected);
		} else {
			assertThatThrownBy(
				() -> testDemo.addPositive(a, b))
					.isInstanceOf(IllegalArgumentException.class);
		}
	}
	
	static Stream<Arguments> argumentsForAddPositive() {
		return Stream.of(
			arguments(2,4,6,false),
			arguments(-2,4,2,true),
			arguments(2,0,2,true),
			arguments(0,-4,-4,true));
	}

	@Test
	void assertThatPairsOfPositiveNumbersAreAddedCorrectly() {
		assertThat(testDemo.addPositive(4, 5)).isEqualTo(9);
		assertThat(testDemo.addPositive(17, 19)).isEqualTo(36);
	}
	
	@ParameterizedTest
	@MethodSource("com.promineotech.TestDemoUnitTest#argumentsForPrimeCheck")
	void assertThatNumberIsCheckedToBePrime(int n, boolean expected) {
		assertThat(testDemo.isNumPrime(n)).isEqualTo(expected);
	}
	
	static Stream<Arguments> argumentsForPrimeCheck(){
		return Stream.of(
				arguments(7,true),
				arguments(8,false),
				arguments(1,false),
				arguments(0,false),
				arguments(174440041,true),
				arguments(-9737337,false));
	}
	
	@Test
	void assertThatNumberSquaredIsCorrect() {
		TestDemo mockDemo = spy(testDemo);
		doReturn(5).when(mockDemo).getRandomInt();
		
		int fiveSquared = mockDemo.randomNumberSquared();
		
		assertThat(fiveSquared).isEqualTo(25);
	}

}
